pyproject.toml
[project]
name = "anychat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.98.0",
    "streamlit>=1.47.1",
    "requests>=2.31.0",
]
